Title: 계단 오르기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240806/계단 오르기/climbing-stairs.cpp
@@ -5,7 +5,7 @@ int main() {
 	int n;
 	cin >> n; 
 
-	int dp[1001] = { 0 }; 
+	long long dp[1001] = { 0 }; 
 	dp[0] = 0; dp[1] = 0; 
 	dp[2] = 1; 
 	dp[3] = 1;
